+++ FNFTpy/fnft_nsev_wrapper.py
@@ -28,7 +28,7 @@ Christoph Mahnke, Shrinivas Chimmalgi 2018, 2019
 """
 
 from .typesdef import *
-from .auxiliary import get_lib_path, check_return_code
+from .auxiliary import get_lib_path, check_return_code, get_winmode_param
 from .options_handling import get_nsev_options
 
 
@@ -185,7 +185,7 @@ def nsev_wrapper(D, q, T1, T2, Xi1, Xi2,
 
     """
 
-    fnft_clib = ctypes.CDLL(get_lib_path(), winmode=0)
+    fnft_clib = ctypes.CDLL(get_lib_path(), winmode = get_winmode_param())
     clib_nsev_func = fnft_clib.fnft_nsev
     clib_nsev_func.restype = ctypes_int
     nsev_D = ctypes_uint(D)